data/ move data/stat to data/ becuse it doesn't make sense
